fix: keep checklist context loader server-only
